# f1api/openapi/api.yml
openapi: "3.0.0"
info:
  title: F1 API
  version: "1.0.0"
  description: |
    ## Paging and Sorting

    The following query paramters can be used to control the paging and sorting of collections.

    ### `pageSize`

    Defines the number of elements per page.

    **Default value:** 25

    **Notice:** maximum number is 100

    ### `pageNumber`

    Specifies the page number that should be returned.

    **Notice:** first page has the number 0

    **Default value:** 0

    ### `sort`

    Specifies by which properties to sort. <br>
    By default the sort order is ascending. This can be changed by adding `asc` for ascending or `desc` for
    descending separated by a comma after the sort property.

    **Hint:** To sort by more than one property, simply add a new sort parameter to the query.

    ### Example

    GET `/drivers?pageSize=20&pageNumber=0&sort=firstName,asc`

    will return the first page with a maximum of 20 drivers sorted by first name in ascending order.
...
paths:
...
  /seasons/{seasonId}:
    get:
      summary: Get season
      tags:
        - "Seasons"
        - "SeasonController"
      parameters:
        - $ref: "./components/parameters.yml#/seasonId"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Season"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        500:
          $ref: "./components/responses.yml#/500"
    patch:
      x-internal: true
      summary: Update season
      tags:
        - "Seasons"
        - "SeasonController"
      parameters:
        - $ref: "./components/parameters.yml#/seasonId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./components/schemas.yml#/Season"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Season"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        409:
          $ref: "./components/responses.yml#/409"
        422:
          $ref: "./components/responses.yml#/422"
        500:
          $ref: "./components/responses.yml#/500"
...
  /teams/{teamId}:
    get:
      summary: Get team
      tags:
        - "Teams"
        - "TeamController"
      parameters:
        - $ref: "./components/parameters.yml#/teamId"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Team"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        500:
          $ref: "./components/responses.yml#/500"
    patch:
      x-internal: true
      summary: Update team
      tags:
        - "Teams"
        - "TeamController"
      parameters:
        - $ref: "./components/parameters.yml#/teamId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./components/schemas.yml#/Team"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Team"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        409:
          $ref: "./components/responses.yml#/409"
        422:
          $ref: "./components/responses.yml#/422"
        500:
          $ref: "./components/responses.yml#/500"
...
  /teamsOfSeasons:
    x-internal: true
    post:
      summary: Add drivers and engine to a team of a season
      tags:
        - "TeamsOfSeasons"
        - "TeamOfSeasonController"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./components/schemas.yml#/TeamOfSeason"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/TeamOfSeason"
        400:
          $ref: "./components/responses.yml#/400"
        409:
          $ref: "./components/responses.yml#/409"
        422:
          $ref: "./components/responses.yml#/422"
        500:
          $ref: "./components/responses.yml#/500"
    get:
      summary: Get teamsOfSeasons / Get teamOfSeason by seasonId and teamId
      tags:
        - "TeamsOfSeasons"
        - "TeamOfSeasonController"
      parameters:
        - $ref: "./components/parameters.yml#/seasonIdQuery"
        - $ref: "./components/parameters.yml#/teamIdQuery"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/TeamsOfSeasons"
        400:
          $ref: "./components/responses.yml#/400"
        500:
          $ref: "./components/responses.yml#/500"
  /teamsOfSeasons/{teamOfSeasonId}:
    x-internal: true
    patch:
      summary: Add driver/drivers to teamOfSeason
      tags:
        - "TeamsOfSeasons"
        - "TeamOfSeasonController"
      parameters:
        - $ref: "./components/parameters.yml#/teamOfSeasonId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./components/schemas.yml#/TeamOfSeasonPatch"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/TeamOfSeason"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        422:
          $ref: "./components/responses.yml#/422"
        500:
          $ref: "./components/responses.yml#/500"
    get:
      summary: Get teamOfSeason / Not particularly needed
      tags:
        - "TeamsOfSeasons"
        - "TeamOfSeasonController"
      parameters:
        - $ref: "./components/parameters.yml#/teamOfSeasonId"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/TeamOfSeason"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        500:
          $ref: "./components/responses.yml#/500"

  # DriverOfRaceController
  #/driversOfRaces:
  #  x-internal: true
  #  get:
  #    summary: Get driversOfRaces / Get driverOfRace by raceId and driverId
  #    tags:
  #      - "DriversOfRaces"
  #      - "DriverOfRaceController"
  #    parameters:
  #      - $ref: "./components/parameters.yml#/raceIdQuery"
  #      - $ref: "./components/parameters.yml#/driverIdQuery"
  #    responses:
  #      200:
  #        description: OK
  #        content:
  #          application/json:
  #            schema:
  #              $ref: "./components/schemas.yml#/DriversOfRaces"
  #      400:
  #        $ref: "./components/responses.yml#/400"
  #      500:
  #        $ref: "./components/responses.yml#/500"
  #/driversOfRaces/{driverOfRaceId}:
  #  x-internal: true
  #  patch:
  #    summary: Update driverOfRace
  #    tags:
  #      - "DriversOfRaces"
  #      - "DriverOfRaceController"
  #    parameters:
  #      - $ref: "./components/parameters.yml#/driverOfRaceId"
  #    requestBody:
  #      required: true
  #      content:
  #        application/json:
  #          schema:
  #            $ref: "./components/schemas.yml#/DriverOfRace"
  #    responses:
  #      200:
  #        description: OK
  #        content:
  #          application/json:
  #            schema:
  #              $ref: "./components/schemas.yml#/DriverOfRace"
  #      400:
  #        $ref: "./components/responses.yml#/400"
  #      404:
  #        $ref: "./components/responses.yml#/404"
  #      409:
  #        $ref: "./components/responses.yml#/409"
  #      422:
  #        $ref: "./components/responses.yml#/422"
  #      500:
  #        $ref: "./components/responses.yml#/500"
  #  get:
  #    summary: Get driverOfRace / Not particularly needed
  #    tags:
  #      - "DriversOfRaces"
  #      - "DriverOfRaceController"
  #    parameters:
  #      - $ref: "./components/parameters.yml#/driverOfRaceId"
  #    responses:
  #      200:
  #        description: OK
  #        content:
  #          application/json:
  #            schema:
  #              $ref: "./components/schemas.yml#/DriverOfRace"
  #      400:
  #        $ref: "./components/responses.yml#/400"
  #      404:
  #        $ref: "./components/responses.yml#/404"
  #      500:
  #        $ref: "./components/responses.yml#/500"